refactor(desktop): organize chat composer ui hooks
