+++ desktop/src/components/workspace/chat/input/ChatInput.tsx
@@ -20,18 +20,18 @@ import {
   useActiveSessionRunningState,
 } from "@/hooks/chat/use-active-chat-session-selectors";
 import { useChatAvailabilityState } from "@/hooks/chat/use-chat-availability-state";
-import { useChatComposerKeyboard } from "@/hooks/chat/use-chat-composer-keyboard";
+import { useChatComposerKeyboard } from "@/hooks/chat/ui/use-chat-composer-keyboard";
 import { useChatDraftState } from "@/hooks/chat/use-chat-draft-state";
 import { useChatModelSelectorState } from "@/hooks/chat/use-chat-model-selector-state";
 import { useChatPromptActions } from "@/hooks/chat/use-chat-prompt-actions";
 import type { PromptAttachmentController } from "@/hooks/chat/use-chat-prompt-attachments";
-import { useComposerSubmitGate } from "@/hooks/chat/use-composer-submit-gate";
+import { useComposerSubmitGate } from "@/hooks/chat/ui/use-composer-submit-gate";
 import { usePlanDraftAttachments } from "@/hooks/plans/facade/use-plan-draft-attachments";
 import { useChatSessionControls } from "@/hooks/chat/use-chat-session-controls";
 import { useQueuedPromptEdit } from "@/hooks/chat/use-queued-prompt-edit";
 import { useActiveReviewRun } from "@/hooks/reviews/facade/use-active-review-run";
 import { useReviewActions } from "@/hooks/reviews/workflows/use-review-actions";
-import { useComposerTextareaAutosize } from "@/hooks/chat/use-composer-textarea-autosize";
+import { useComposerTextareaAutosize } from "@/hooks/chat/ui/use-composer-textarea-autosize";
 import { focusChatInput } from "@/lib/domain/focus-zone";
 import { serializeChatDraftToPrompt } from "@/lib/domain/chat/transcript/file-mentions";
 import { promptAttachmentSnapshotsToContentParts } from "@/lib/domain/chat/composer/prompt-attachment-snapshot";
