+++ desktop/src/components/workspace/chat/input/ModelSelector.tsx
@@ -15,7 +15,7 @@ import {
   ProviderIcon,
   Search,
 } from "@/components/ui/icons";
-import { useModelSelectorMenu } from "@/hooks/chat/use-model-selector-menu";
+import { useModelSelectorMenu } from "@/hooks/chat/ui/use-model-selector-menu";
 import { useNativeOverlayRegistration } from "@/hooks/ui/use-native-overlay-presence";
 import { ComposerControlButton } from "./ComposerControlButton";
 import { PendingConfigIndicator } from "./PendingConfigIndicator";
