+++ desktop/src/hooks/chat/ui/use-chat-composer-keyboard.test.ts
