+++ desktop/src/hooks/chat/ui/use-chat-dock-inset.ts
