+++ desktop/src/hooks/chat/ui/use-composer-dock-slots.tsx
