+++ desktop/src/hooks/chat/ui/use-composer-submit-gate.test.ts
