+++ desktop/src/hooks/chat/ui/use-composer-textarea-autosize.ts
