+++ desktop/src/hooks/chat/ui/use-model-selector-menu.ts
